extract buffer access and stream reader methods
